# backend/api/fenzvideo/v1/video_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v6.33.4
// source: fenzvideo/v1/video.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationVideoServiceCreateVideo = "/fenzvideo.v1.VideoService/CreateVideo"
const OperationVideoServiceDeleteVideo = "/fenzvideo.v1.VideoService/DeleteVideo"
const OperationVideoServiceGetRecommended = "/fenzvideo.v1.VideoService/GetRecommended"
const OperationVideoServiceGetVideo = "/fenzvideo.v1.VideoService/GetVideo"
const OperationVideoServiceTogglePublish = "/fenzvideo.v1.VideoService/TogglePublish"
const OperationVideoServiceUpdateVideo = "/fenzvideo.v1.VideoService/UpdateVideo"

type VideoServiceHTTPServer interface {
	CreateVideo(context.Context, *CreateVideoRequest) (*VideoReply, error)
	DeleteVideo(context.Context, *DeleteVideoRequest) (*DeleteVideoReply, error)
	GetRecommended(context.Context, *GetRecommendedRequest) (*VideoListReply, error)
	GetVideo(context.Context, *GetVideoRequest) (*VideoReply, error)
	TogglePublish(context.Context, *TogglePublishRequest) (*VideoReply, error)
	UpdateVideo(context.Context, *UpdateVideoRequest) (*VideoReply, error)
}

func RegisterVideoServiceHTTPServer(s *http.Server, srv VideoServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/videos", _VideoService_CreateVideo0_HTTP_Handler(srv))
	r.GET("/api/v1/videos/{id}", _VideoService_GetVideo0_HTTP_Handler(srv))
	r.PUT("/api/v1/videos/{id}", _VideoService_UpdateVideo0_HTTP_Handler(srv))
	r.DELETE("/api/v1/videos/{id}", _VideoService_DeleteVideo0_HTTP_Handler(srv))
	r.PATCH("/api/v1/videos/{id}/publish", _VideoService_TogglePublish0_HTTP_Handler(srv))
	r.GET("/api/v1/videos/recommended", _VideoService_GetRecommended0_HTTP_Handler(srv))
}

func _VideoService_CreateVideo0_HTTP_Handler(srv VideoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateVideoRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationVideoServiceCreateVideo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateVideo(ctx, req.(*CreateVideoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*VideoReply)
		return ctx.Result(200, reply)
	}
}

func _VideoService_GetVideo0_HTTP_Handler(srv VideoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetVideoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationVideoServiceGetVideo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetVideo(ctx, req.(*GetVideoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*VideoReply)
		return ctx.Result(200, reply)
	}
}

func _VideoService_UpdateVideo0_HTTP_Handler(srv VideoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateVideoRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationVideoServiceUpdateVideo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateVideo(ctx, req.(*UpdateVideoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*VideoReply)
		return ctx.Result(200, reply)
	}
}

func _VideoService_DeleteVideo0_HTTP_Handler(srv VideoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteVideoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationVideoServiceDeleteVideo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteVideo(ctx, req.(*DeleteVideoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteVideoReply)
		return ctx.Result(200, reply)
	}
}

func _VideoService_TogglePublish0_HTTP_Handler(srv VideoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in TogglePublishRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationVideoServiceTogglePublish)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TogglePublish(ctx, req.(*TogglePublishRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*VideoReply)
		return ctx.Result(200, reply)
	}
}

func _VideoService_GetRecommended0_HTTP_Handler(srv VideoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetRecommendedRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationVideoServiceGetRecommended)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetRecommended(ctx, req.(*GetRecommendedRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*VideoListReply)
		return ctx.Result(200, reply)
	}
}

type VideoServiceHTTPClient interface {
	CreateVideo(ctx context.Context, req *CreateVideoRequest, opts ...http.CallOption) (rsp *VideoReply, err error)
	DeleteVideo(ctx context.Context, req *DeleteVideoRequest, opts ...http.CallOption) (rsp *DeleteVideoReply, err error)
	GetRecommended(ctx context.Context, req *GetRecommendedRequest, opts ...http.CallOption) (rsp *VideoListReply, err error)
	GetVideo(ctx context.Context, req *GetVideoRequest, opts ...http.CallOption) (rsp *VideoReply, err error)
	TogglePublish(ctx context.Context, req *TogglePublishRequest, opts ...http.CallOption) (rsp *VideoReply, err error)
	UpdateVideo(ctx context.Context, req *UpdateVideoRequest, opts ...http.CallOption) (rsp *VideoReply, err error)
}

type VideoServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewVideoServiceHTTPClient(client *http.Client) VideoServiceHTTPClient {
	return &VideoServiceHTTPClientImpl{client}
}

func (c *VideoServiceHTTPClientImpl) CreateVideo(ctx context.Context, in *CreateVideoRequest, opts ...http.CallOption) (*VideoReply, error) {
	var out VideoReply
	pattern := "/api/v1/videos"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationVideoServiceCreateVideo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *VideoServiceHTTPClientImpl) DeleteVideo(ctx context.Context, in *DeleteVideoRequest, opts ...http.CallOption) (*DeleteVideoReply, error) {
	var out DeleteVideoReply
	pattern := "/api/v1/videos/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationVideoServiceDeleteVideo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *VideoServiceHTTPClientImpl) GetRecommended(ctx context.Context, in *GetRecommendedRequest, opts ...http.CallOption) (*VideoListReply, error) {
	var out VideoListReply
	pattern := "/api/v1/videos/recommended"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationVideoServiceGetRecommended))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *VideoServiceHTTPClientImpl) GetVideo(ctx context.Context, in *GetVideoRequest, opts ...http.CallOption) (*VideoReply, error) {
	var out VideoReply
	pattern := "/api/v1/videos/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationVideoServiceGetVideo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *VideoServiceHTTPClientImpl) TogglePublish(ctx context.Context, in *TogglePublishRequest, opts ...http.CallOption) (*VideoReply, error) {
	var out VideoReply
	pattern := "/api/v1/videos/{id}/publish"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationVideoServiceTogglePublish))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *VideoServiceHTTPClientImpl) UpdateVideo(ctx context.Context, in *UpdateVideoRequest, opts ...http.CallOption) (*VideoReply, error) {
	var out VideoReply
	pattern := "/api/v1/videos/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationVideoServiceUpdateVideo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}