# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/videos:
        post:
            tags:
                - VideoService
            operationId: VideoService_CreateVideo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.CreateVideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.VideoReply'
    /api/v1/videos/recommended:
        get:
            tags:
                - VideoService
            operationId: VideoService_GetRecommended
            parameters:
                - name: sessionId
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.VideoListReply'
    /api/v1/videos/{id}:
        get:
            tags:
                - VideoService
            operationId: VideoService_GetVideo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.VideoReply'
        put:
            tags:
                - VideoService
            operationId: VideoService_UpdateVideo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.UpdateVideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.VideoReply'
        delete:
            tags:
                - VideoService
            operationId: VideoService_DeleteVideo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.DeleteVideoReply'
    /api/v1/videos/{id}/publish:
        patch:
            tags:
                - VideoService
            operationId: VideoService_TogglePublish
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.TogglePublishRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.VideoReply'
components:
    schemas:
...
        fenzvideo.v1.CreateVideoRequest:
            type: object
            properties:
                title:
                    type: string
                description:
                    type: string
                categoryId:
                    type: string
                tagIds:
                    type: array
                    items:
                        type: string
                accessTier:
                    type: integer
                    format: int32
                duration:
                    type: integer
                    format: uint32
                videoUrl:
                    type: string
                thumbnailUrl:
                    type: string
        fenzvideo.v1.DeleteVideoReply:
            type: object
            properties:
                success:
                    type: boolean
...
        fenzvideo.v1.TagItem:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                slug:
                    type: string
...
        fenzvideo.v1.TogglePublishRequest:
            type: object
            properties:
                id:
                    type: string
                isPublished:
                    type: boolean
...
        fenzvideo.v1.UpdateVideoRequest:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                description:
                    type: string
                categoryId:
                    type: string
                tagIds:
                    type: array
                    items:
                        type: string
                accessTier:
                    type: integer
                    format: int32
                thumbnailUrl:
                    type: string
        fenzvideo.v1.VideoListReply:
            type: object
            properties:
                videos:
                    type: array
                    items:
                        $ref: '#/components/schemas/fenzvideo.v1.VideoReply'
                total:
                    type: string
        fenzvideo.v1.VideoReply:
            type: object
            properties:
                id:
                    type: string
                userId:
                    type: string
                username:
                    type: string
                categoryId:
                    type: string
                categoryName:
                    type: string
                title:
                    type: string
                description:
                    type: string
                videoUrl:
                    type: string
                thumbnailUrl:
                    type: string
                duration:
                    type: integer
                    format: uint32
                views:
                    type: string
                accessTier:
                    type: integer
                    format: int32
                isPublished:
                    type: boolean
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/fenzvideo.v1.TagItem'
                createdAt:
                    type: string
...